remove legacy edge-api module
